Fix odd behavior of double-click in register
If a user is viewing the ledger, and a transaction is selected (such as
if it was just edited it) and then a double click is performed on
another transaction, it edits the previously selected transaction but
not the one that was clicked on. This only happens if the first
transaction is reconciled and a dialog pops up before the transaction is
edited. This change fixes the behavior.

BUG: 397457
FIXED-IN: 5.0.2