Don't hardcode custom icons location
Custom icons aren't visible on MS Windows because path to them is set to
install directory. It happens to be that this install directory isn't
the directory where KMyMoney searches for icons on MS Windows. This
patch intends to search for the icons before it adds them to icons
search path. It's untested though.

Differential Revision: https://phabricator.kde.org/D15087