Work-around for missing statement end date
This change amends commit 3e95b8c4ec4b6. Some XML elements of an OFX
file/download are not processed by LibOFX. Among them is the statement
date information for investment statements.

This adds a work-around which uses the post date of the youngest
transaction as the statement end date should the real thing be missing.

BUG: 343878
FIXED-IN: 5.0.2