Fix placement of calculator
The calculator is placed in the middle of the screen when it is first
opened after transaction edit started. If closed and re-opened in the
same transaction edit session it is placed correctly. This change places
the calculator always in the right position next to the amount edit
widget.

Fix taken from commit 6a221524755177087fb1dcd110af1bc5f592f61a

CCBUG: 283784