Fix QIF import file selection dialog
When selecting a file to import the file selection screen defaults to
"Import Files" and does not show ".QIF" files. It does show ".qif"
files....lower case. As a result, user has to switch to "All Files"
view.

In addition, when navigating to the directory location of the file to
import.....when importing another file.....the previous location is not
remembered. User has to navigate to the location again.

This change fixes the above mentioned problems. For existing user
defined profiles, the filter may be amended to contain "*.qif *.QIF" to
show lower and uppercase filenames using the QIF import plugin settings
dialog.

BUG: 399673