Update the name of the OFX plugin to its new name
Summary:
Up to and including 4.8 the OFX importer plugin was called "KMyMoney
OFX". From version 5 on it is called "ofximporter". So we update it here
in case we find the 'old' name.

This fixes the XML and SQL backend but I am sure it can be done in a more
elegant fashion using SQL statements.

Reviewers: ostroffjh

Differential Revision: https://phabricator.kde.org/D15483