Add support for saving and loading tax account mappings to and from a template file
A tax account mapping is stored in an xml tag of type 'flag' with
the attribute name='VatAccount'. The attribute 'value' contains a
reference to the corresponding tax account, which contains the
reference in the attribute 'id'.

When importing a template file with a tax account assignment, the
reference is first stored in a temporary key/value pair with the
name 'UnresolvedVatAccount' of the account concerned and after
completion of the load is stored in the key/value pair 'VatAccount'
based on the ID of the tax account now available.

CCBUG:380286
Reviewed By: tbaumgart
Differential Revision: https://phabricator.kde.org/D16726