Make sure to display full title on charts
Upstream library software (kdiagram) causes the title of a chart to be
chopped off on the right side as shown in attachments to bug #399260. I
used gammaray to see that the full title name was passed to the drawing
routines inside kdiagram but still it got truncated. Adding a few extra
spaces (5 in this case) at the end solved the issue.

BUG: 399260
FIXED-IN: 5.0.3