Unselect item in case text field is empty during focus out
In case the combo box contained a previously selected item, this was not
reset when the combo box text field was empty and the focus was removed
from the field. This caused the previously selected item still to be
reported as current item during e.g. the storage operation of a
transaction.

BUG: 402120
FIXED-IN: 5.0.3