Correctly read opposite signs option from banking profile
The "Opposite signs" checkbox is not being initialized from the value
stored in the Banking profile in csvimporterrc. The value is stored
correctly in csvimporterrc as OppositeSigns=true but the checkbox is
always unchecked when the profile is re-used. This change fixes the
problem.

Thanks to Mark for providing the fix.

BUG: 402195
FIXED-IN: 5.0.3