Don't post superfluous question
In case the creation of a new transaction is started but no data was
entered and the user clicks on an area outside of the transaction editor
a confirmation message popped up asking if he wants to save the data
(when there was none). This has now been modified so that the editor is
silently closed in case the 'Enter' button is disabled.

As a side effect, one of the questions was adjusted to reflect the
actual situation.

BUG: 402302
FIXED-IN: 5.0.3