Automatically adjust tax split if possible
Importing transaction assigns categories based on previous transactions
found with the same payee or a preselected category. In case such a
historic transaction contains multiple splits then these splits are
copied but the amounts of the splits are not adjusted based on the value
of the imported transaction.

This change now adjusts in case the resulting transaction consists of an
asset/liability account, a taxable category and a tax category such that
the sum of all splits is zero based on the imported transaction.

This also fixes the "Copy splits" feature when used on a taxed
transaction.

BUG: 402794
FIXED-IN: 5.0.3