Fix customized report loading and saving
The storage of the accountgroup information for customized reports was
broken due to a mis-assignment between numeric and textual
representation. This was introduced with
https://phabricator.kde.org/D13581 which landed as commit
b7c44b7f705cd8.

With the current change, the account group information will be
reconstructed during loading of the file depending on the report type
and the above mentioned assignment problem will be fixed.

Also, the report type was not written for any customized information
report which caused a crash during usage after the next start of the
application. Since the report type cannot be detected automatically from
the available data in broken files, the type will be set to AccountInfo.

A respective message is dumped to the console. Users may have to
redesign their loan, schedule and schedule summary reports. The problem
shows, if a user opens such a report and it now displays an account
information report. Nevertheless, the original report filter settings
are loaded correctly.

All other report types are not affected by this problem.

BUG: 403068
FIXED-IN: 5.0.3