Fix security creation logic
In case no smallest account fraction is provided in the ctor of a
MyMoneySecurity object it should use the same value as provided for the
smallest cash fraction. Providing a default value of 100 destroys this
logic. This forced Palladium to have values always rounded to two
decimals.

In case the user has such a security on file there is no other way as
manually fixing it by copying the value from the SCF attribute to the
SAF attribute in the XML file and the value in the smallestCashFraction
column to the smallestAccountFraction column in the database.

CCBUG: 403249