Setup the transaction commodity based on the trading currency
For new investment transactions the commodity was always set to the
baseCurrency instead of the securities trading currency. This usually
caused the values to be rounded to two decimals (smallest account
fraction of the base currency).

This change fixes the problem by not looking at the empty original
transaction in this case.

BUG: 403249
FIXED-IN: 5.0.3