Improve transaction importer
The transaction importer - used during online banking statement download
or file import - used to either copy the categorization of a transaction
that exactly matched the value of the imported transaction or the last
transaction found for that payee in the same account.

This is a bit crazy if you have two alternating and slightly varying
payments from/to the same payee with very different amouts. This leads
to the fact, that in most cases it will choose the last one (which is
the wrong one in this scenario).

With this change the importer now looks at all historical transactions
for the payee in the account and if no exact match is found use the
transaction with the smallest difference in value to the imported
transaction. This improves the above scenario very much.