Fix plugin loader
The Qt plugin system searches and reports plugins in the following
order:

- current directory
- paths provided in QT_PLUGIN_PATH (separated by QDir::listSeparator)
- path provided by QLibraryInfo::location(QLibraryInfo::PluginsPath)

The KMyMoney plugin loader stores all plugin metadata of the returned
plugins in a QMap to process them. It was not checked if a plugin was
already found and so the last instance always won the game leaving the
ones in the current directory or in one of the directories mentioned in
QT_PLUGIN_PATH without effect.

This change makes sure to load and use the first plugin found and also
prints some more information on the console which plugins were found and
which ones are loaded.