Improve result information for OFX mapping
The source used to obtain information necessary for OFX mapping
(www.ofxhome.com) also presents information about the availability of
the provided URL and the verification of the SSL/TLS certificate.

Since this information was not interpreted by KMyMoney, URLs were
presented that are known to be broken.

This change shows an error message when a non working URL or an invalid
certificate is found on ofxhome.com.

GUI: