Fix check number handling
KMyMoney versions prior to 5.0 allowed to re-use check numbers. This is
not possible with the current 5.0 versions. This change will bring back
the original functionality.

In case KMyMoney detects a duplicate check number it informs the user.
The user then has the possibility to assign the next free number or
leave the number as entered. It will not be cleared in this case
anymore.

BUG: 402101
FIXED-IN: 5.0.3