Support sell w/o asset account
Summary:
Allow a sell operation that reduces the number of shares to pay fees.
This is to handle mutual fund account fees. There is no money
transferred -- only a reduction in shares.

BUG: 398133

Test Plan: I have tested entry and retrieval of the transaction. What needs to be tested how this change affects reporting. I can't do that.

Reviewers: ostroffjh

Differential Revision: https://phabricator.kde.org/D15215