Do not allow to enter transactions prior to account opening date
The transaction editor in the standard account ledger does not allow to
enter transactions that reference accounts that have an opening date
later than the transaction date.

The investment transaction editor is missing this feature and one could
enter transactions with a date prior to the opening date of the
investment or asset account referenced in the transaction.

Moreover, the opening date of an invest account cannot be changed by the
user but got updated to a matching value when the consistency check was
run (e.g. as part of file save). The user was informed and was
surprised.

This change adds an automatic update of the opening date for the invest
account if needed when the transaction is entered. Also, it does not
allow to enter the transaction, colorizes the date and presents a
tooltip on the date widget just like for the standard transaction
editor.

BUG: 403886
FIXED-IN: 5.0.4