Added automatic fix to adjust settings of precious metals
Files created before Jan 19th 2019 will most certainly have a false
setting for the smallest account fraction of the precious metals gold,
platinum, palladium and silver.

This change fixes the problem in the file when the file is opened.

At the same time it initializes an investment transaction with the
commodity of the account and not the base currency.

BUG: 403885
FIXED-IN: 5.0.4