Improve memo column selection in CSV import wizard
The CSV import wizard did allow to select multiple columns for memo
fields but it was not visible to the user which ones were selected.
Also, there is no way to clear the memo selection only but only all
column assignments.

This change adds a display item that shows the selected columns and also
adds a clear button to clear the memo field selection only.

GUI:
BUG: 404156
FIXED-IN: 5.0.4