Improve date change feature
The current behavior of the + and - key in a date selection widget was
to increase or decrease by day and roll over into the next resp. previous
month. This has been enhanced to increase/decrease by day/month/year
depending on the current section holding the cursor with the month doing
a roll over into the next/previous year.

BUG: 399681