Add support for assigning tags in transaction splits editor
In the transaction split editor the tags column is designed similar
to categories and shows the assigned tags as text.
On editing a split a drop down box with all available tags and
checkboxes for removing assigned tags are visible.

Because tags are always assigned to splits, adding tags by the transaction
split editor works in the same way as adding tags in the ledger view.

Please note that the report engine is not able to filter
tags on categories yet.

Ported to KF5 from commit c712d81154ef2fd5bb136bc4521ba3d046b74ba6.

BUG:341589
FIXED-IN:4.8.4,5.0.4
Reviewed By: tbaumgart
Differential Revision: https://phabricator.kde.org/D18789