Forecast (history) has no effect on home view
The home view section selector in the settings has an entry "Forecast
(history)" which does not show any effect on screen when changed. Since
it is currently not used, it is removed from the settings list. It will
be kept in source code (marked as unused) since the order of the entries
and their index must not change to maintain backward compatibility of
the settings.

BUG: 406074
FIXED-IN: 5.0.4