Provide method to construct unique transaction ID for OFX import
User reports show, that the FITID attribute provided by some
institutions during OFX statement download is not as unique as it is
specified. In fact, some banks issue a different FITID every time you
download the same statement.

This makes the FITID attribute absolutely worthless for duplicate
detection.

Since asking banks to fix their software has no reasonable chance to
lead to success on short call, this change allows to switch the
duplicate detection to be based on a method provided by KMyMoney. In
fact it uses the same algorithm which works well in the KBanking
importer for years already. Note: switching from "OFX FITID" to
"KMyMoney Hash" may result in duplicates for one more time.

At the same time, the OFX import options are now available during file
import and not only for mapped accounts (OFX direct connect).

BUG: 390681
FIXED-IN: 5.0.4