doc: drop almost all bits related to installation
Summary:
In the majority of the cases, users install kmymoney as available in
their distribution, and thus most of this content is not useful; users
that want to compile kmymoney from sources find already instructions on
kmymoney.org, in the development wikis, and in the in-source files, so
no need to duplicate them in a user-taylored documentation.

Furthermore, the current building instructions are outdated, and an
updated version, together with any FAQ related to building from sources
and developing on kmymoney, belongs to online resources (easier to keep
up-to-date).

Test Plan:
- build the documentation
- verify the mentioned bits are not present

Reviewers: ostroffjh

Reviewed By: ostroffjh

Subscribers: tbaumgart, #kmymoney, kde-doc-english

Tags: #documentation

Differential Revision: https://phabricator.kde.org/D20389