Fix subtotals
Subtotals in e.g transaction reports are not correctly calculated and
shown. This only shows up if a the category type (e.g. income, expense)
and the sub-category name change and the top-level category name is the
same.

This change fixes the problem so that all sub-totals are calculated and
presented correctly.

BUG: 406525
FIXED-IN: 5.0.4