Show all asset and liability accounts during statement import
This change is a workaround to allow access to all asset and liability
accounts despite the information provided in the data from the
institution (which we have learned in some cases is just wrong).

A better approach would be to have a UI option to enhance the list does
not contain the account the user is looking for.

BUG: 396225
BUG: 392305
FIXED-IN: 5.0.4