Provide internal OFX id generation also for WebConnect
The change provided in commit 65e3837697c93786 is not available when
importing an OFX statement file via WebConnect. For this case, a new
option has been added (no UI yet) to allow selection of the default for
the unique transaction id to be either the OFX FITID field or our
internal hash. The default is to use OFX FITID, so no change in prior
behavior. This default settings will also be used to preset the combo
box in the file selection dialog for OFX import and the account setting
for mapped accounts.

If users set this option, they need to know that some duplicate
transactions may appear since the method to detect them changed. One way
to deal with that is to remove both transactions and redo the import.

To set the method to the internal hash create a file named

  ~/.config/kmymoney/ofximporterrc

with the following content:

  [General]
  useOwnFITID=true

Future versions of KMyMoney will provide a UI to change this setting.

CCBUG: 390681