Make XML files comparable again
Since commit e4d384d34a91eb6024800 (which later made it into v5.0.2),
the order of attributes in the XML file is random each time the file is
saved. This leads to the fact that it is extremly hard to compare two
files because one does not see the forest for the trees.

This change removes the randomness and thus brings back comparability.

BUG: 396978
FIXED-IN: 5.0.4