Reset column assignment when out of bounds
In case the column assignment is cleared using the clear all button, the
profile is not updated because there is no change in index in combobox
widget and the connected slot for profile update will never be called.

This change makes sure that in such a case (previous assignment is out
of bounds for the current file) the assignment is reset before it is
used. This avoids the out-of-bounds crash later on.

BUG: 408026
FIXED-IN: 5.0.5