Update payee in matched transaction
In case two payees are merged and one of them is referenced in a matched
transaction, it could happen that the transaction was removed from the
users data leading to incorrect data.

This change also updates the payee identifier in the matched transaction
should it exist and thus avoids data loss.

Deleting a payee that is referenced in a matched transaction still is
not possible but does not result in data loss.

BUG: 408205
FIXED-IN: 5.0.5