Select amount when changing focus using the tab/backtab key
The previous commit showed a problem in the transaction editor such that
the amount was not selected when the focus was changed to the amount
field using the tab or backtab key. This used to work before with the
KMyMoneyEdit object.

This change fixes the problem by passing the reason as part of the focus
change.