Provide more descriptive text for holiday region
The holiday region selection showed the same text for different entries
in the same country/language region. For Germany alone there were 14
entries with the same text so one could not differentiate between them.

This change fills the combo box with the description of the holiday
region which allows to differentiate between those entries if it is
available. If it is empty, the current more general text is used.