Use mapped account only if information is present
Using the OFX (file) importer may not provide information about the
account the imported data relates to. In this case, it does not make
sense to search for an online mapping. Doing so leads to false results,
as without values for account or routing number the first mapped account
will be selected which is wrong.

This change does not try the mapping if both values are empty and allows
the user to select the account as part of the process.

CCBUG: 405206