Improve handling of decimal point character in amount edits
Pressing the decimal point in an amount edit field with all text
selected caused the selection to be removed and the cursor to jump to
the end of the amount.

This change now forces the text to be replaced with a 0 so that pressing
the decimal point with all text selected allows to start entry of a
fractional amount without pressing a leading 0.

BUG: 409371