Select amount when changing focus using the tab/backtab key
Commit 8841ae39d15b5ddb916a151d7bc20415862fe34d introduced a problem
such that the amount was not selected when the focus was changed to the
amountfield using the tab or backtab key. This used to work before with
the KMyMoneyEdit object.

For the standard transaction editor this was fixed by commit
5008fc800e8806522263a9c9b5cc1cc5d74c7ae5 but there are more locations in
the code where this change is necessary.

This change adds the fix to cover all locations which show the above
problem.

BUG: 409371