Raise minimum required version of AqBanking
the German online banking users have to use a registered software from
mid-September onwards and the institutions are allowed according to EU
regulations to reject/refuse access by non-registered client
applications.

To be able to support this process KMyMoney requires a newer version of
AqBanking which has been released lately.