Do not move all splits of a transaction
In KMyMoney, in a transaction with multiple splits, where more than one
split is assigned to the same category, attempts to change only one
split to a new category changes all splits with the same category to the
new category.

This change only moves the selected transaction. Also, it only removes
those accounts from the list that are not referenced by the selected
splits which was detected during the course of the fix of the original
problem.

BUG: 411030
FIXED-IN: 5.0.6