Adding additional variable to check printing
Summary: When using check forms that contain a log for record keeping it is
needed to print out all of the splits associated with a transaction when
printing a check. This cannot be done since the split variables are not
present in the check printing plugin.

This patch will add the variables SplitAccountName and SplitValue for
up to 11 splits within a transaction. The check template will then be
able to access the split variables and print them when printing a check.

Reviewers: #kmymoney, tbaumgart

Reviewed By: #kmymoney, tbaumgart

Subscribers: tbaumgart

Tags: #kmymoney

Differential Revision: https://phabricator.kde.org/D23369