KBanking differentiates between account types
In case an institution uses the same account number for e.g. for
sub-accounts and investment depots KMyMoney was not able to
differentiate between the two accounts.

This change adds the type to the value used for finding the right
account in the KMyMoney account list.

Also, information about investments will be added to the statement.

Original author: Holger Magnussen <holger@muscate-magnussen.de>