Add an option to use both merge notes and memo when importing GnuCash file, and merge them if both are present
Summary:
I have beeing using GnuCash for 8 years before deciding to move on to kmymoney.
By now, my non-split transactions tend to use either the Memo and Notes fields, depending
on the source they were imported from. So instead of having to choose either of them and
loosing some potentially meaningful information, I added an option to the importer that uses
both of the fields and combines them together if both are present. User now has a choice to
use Notes, Memo, or both.

Differential Revision: https://phabricator.kde.org/D25804