printcheck plugin: Use value instead of shares
from Thomas Baumgart:
Each split has two attributes: shares and value. Each transaction is
denominated in a currency (transaction.currency()). The sum of the
value attributes over all splits of a transaction must be zero for
the transaction to be balanced (that is what double entry accounting
is all about), where it does not matter in which currency that happens.
The shares attribute represents the units in the currency of the account
that the split references. So the sum of all shares over all splits of
a specific account provides the balance of the account.

In the end, it depends on how you want the different currencies printed
on the check. All this is only important in case one has multiple
currencies. I just want to avoid that we add some code that works in
single currency environments but gives the user a headache when using
it with multiple currencies.

see https://phabricator.kde.org/D23369#517852