Adding additional variable to check printing
When using check forms that contain a log for record keeping it is
needed to print out all of the splits associated with a transaction when
printing a check. This cannot be done since the split variables are not
present in the check printing plugin.

This patch will add the variables SplitAccountName and SplitValue for
up to 10 splits within a transaction. The check template will then be
able to access the split variables and print them when printing a check.

Differential Revision: https://phabricator.kde.org/D23369
BUG:415411
FIXED-IN:4.8.5,5.0.7