Fix detection of LibOFX clientuid support
Commit 487f8758987edb72e616a602706f100cafa877d8 introduced a different
method to detect the presence of the LibOFX library. This change also
changed the variable name for the libofx include directory. This caused
the detection of the clientuid to fail later in the process as it still
used the old variable (which now is empty).

This change modifies the variable name and thus fixes the clientuid
detection.