Allow local script files as price update sources in AppImage
The price update service in KMyMoney allows to start local scripts to
fetch price information. This did not work when started as AppImage.
This change brings that functionality to the AppImage version.

Within the AppImage version a configured URL such as

 "file:///home/user/bin/price.sh %1"

is called as "/bin/sh" "-c" "/home/user/bin/price.sh %1" where %1 will
be replaced with the actual investment symbol.

BUG: 418823
FIXED-IN: 5.0.9