Use the correct institution in reports
When an account has no institution assigned the one of the top-level
account was used in some reports. This leads to wrong results in case of
investments which don't have an institution assigned but should use the
one of the parent investment account.

This worked as long as the investment account itself was the top-level
account. In case the investment account itself was a sub-account to a
top-level account or even further down in the account hierarchy, the
institution was always taken from the top-level account.

This change fixes the problem by walking up the hierarchy and using the
first institution found.

BUG: 421260
FIXED-IN: 5.0.9