Fix Account context menu's Reconcile option opens incorrect ledger
Summary:
Selecting the 'Reconcile...' option from account's context menu in the
Accounts view would open ledger of the last account viewed in the
Ledgers view, not of the account selected in the Accounts view.

This commit fixes the issue so that the 'Reconcile...' option opens
correct ledger.

BUG: 421307
FIXED-IN: 5.0.9

Reviewers: tbaumgart

Reviewed By: tbaumgart

Tags: #kmymoney

Differential Revision: https://phabricator.kde.org/D29614