Fix new account wizard not asking if the user wants to add a new payee
When the user entered a new payee name in the New Account Wizard and
then changed focus to another form field by clicking the mouse, the user
was not asked if he or she wanted to add the new payee. The pop-up
question window appeared correctly if focus was changed by keyboard
(e.g. by pressing the Tab key).

This commit fixes the issue so that the user is asked about adding the
new payee when focus is lost due to a mouse click event.

BUG: 421691
FIXED-IN: 5.0.9

(cherry picked from commit 6e1746e3d6ea238b435490881c6241c2a5bcfe18)