Update start date of scheduled transactions
In case an end date was specified for a scheduled transaction and the
next due date was changed by the user after the scheduled transaction
was created, the date of the following transactions was not updated.

This change makes sure that the following transactions are also shown
and executed based on the updated date information.