Update start date of scheduled transactions
In case an end date was specified for a scheduled transaction and the
next due date was changed by the user after the scheduled transaction
was created, the date of the following transactions was not updated.

This change makes sure that the following transactions are also shown
and executed based on the updated date information.

BUG: 421750
FIXED-IN: 5.0.9

(cherry picked from commit fdb3b3ba1b47ca8345c04fd24a79af6ee9c8cdd0)