Fix incorrect account hierarchy if the account is marked as preferred
When an account was marked as preferred and the user edited it, the
account was moved in the hierarchy tree even though the user did not
enter the Hierarchy tab to change anything. When the user opened the
Hierarchy tab, there was Favorites category displayed but it should not.

This commit solves the issue so that a preferred account is not moved
when the user does not want it to. The Favorites catergory is now not
displayed in the Hierarchy tab.

BUG: 422012
FIXED-IN: 5.0.9

Signed-off-by: Robert Szczesiak <dev.rszczesiak@gmail.com>
(cherry picked from commit 5412d8e32c85485f047127fb2b0e2edfff53ddb4)