Set the minimum unit of KRW into 1
Although the subdivision of South Korean Won exists de jure, it is de
facto not used in bank statements and stock price tickers. Only in
foreign exchange markets the subdivision is used but they are rounded
down when it becomes to accounting. Thus adjust the precision of South
Korean Won to 1.

BUG: 420683
FIXED-IN: 5.0.9

(cherry picked from commit 8025e549001ebbdbe3f2aa09074dd5a23ed06836)