Fix ui version information
New xmlgui menu item entries have been added with commit
78c03decc0296d6464fbb74718992919a13952d4 but the version of the file was
not increased. This causes the new menu items to remain hidden when a
user has a modified local version of the file as a result of local
setting modifications (toolbar, shortcuts, etc.).

Increasing the version number should solves the problem.