Fix Budgets view continually asking if the user want to save changes
When the user made changes to a budget and clicked on the update button,
a pop-up windows asked if he or she wants to save the changes. Clicking
'Yes' caused the window to pop up again, and again.

This commit solves the issue so that the pop-up window appears once and
respects the user's choice.

Signed-off-by: Robert Szczesiak <dev.rszczesiak@gmail.com>