Fix keypad comma/dot to use QLocale::decimalPoint for amounts
In general, pressing the keypad's comma or dot character creates the
correct key press event for a locale. In case the keyboard option
"Four-level key with abstract separators" is set in the KDE settings,
the wrong character is used to create the event (e.g.: for locales DE or
FR a dot is returned when the decimal symbol is a comma).

This change fixes the problem such that a key press event for dot or
comma with the Qt::KeypadModifier set is turned into a key press event
with for the QLocale::decimalPoint() character.

BUG: 423509
FIXED-IN: 5.1.1