Harmonize display of closed accounts in home view
Different sections of the home view displayed closed accounts depending
of the "Show all accounts" and "Do not show closed accounts"
differently. The handling has been harmonized such that closed accounts
will be displayed in all sections alike either if "Show all accounts" is
active or "Don't show closed accounts" is inactive.

Closed accounts will also be shown in strike-through so that they can
easily be identified as closed.

BUG: 424098
FIXED-IN: 5.1.1