Add missing release event in very specific case
The date picker object is used in other areas of KMyMoney (e.g.
transaction editor) where this problem does not show up. I analyzed it,
and the problem is, that the very first MouseReleaseButton event on the
date picker after object creation is not reaching the object under
unknown circumstances in the enter schedule dialog. This is not always,
but the scenario reported in the bug entry shows it. I have no idea
about the root cause, though.

Detecting the specific scenario and sending a mouse release button event
based on the data of the press event solves the problem. Drawback:
pressing the mouse long in that scenario now has no effect. It needs a
second press to get the number increasing while the mouse button is
pressed.

BUG: 424378
FIXED-IN: 5.1.1