Allow to calculate one parameter of a loan
When leaving an amount edit field in the new loan account wizard it will
automatically be filled with 0. This will prevent the loan calculator to
identify the field that should be calculated.

Enable the option to allow an empty field prevents the automatic fill
with the value 0 on focusOut event.

BUG: 424511

(cherry picked from commit 19e9ac500810c9026c83a532a7e6e990fd39ea63)