Report all matching transactions in transaction report
In case the first processed split of a multi split transactoin matches
the filter criteria and no other one does, then the transaction was not
included in the report at all.

Since other splits may match the filter also, the first one found will
be stacked and only added, if no other data was added for that
transaction.

BUG: 424674
FIXED-IN: 5.1.1