Show correct currency in report
In multi-currency transactions the transaction commodity was used as
currency instead of the categories currency. The value shown is correct.

This change fixes the problem by overriding the currency in this case
with the one of the category.

BUG: 423870
FIXED-IN: 5.1.1

(cherry picked from commit 49dcabca8bd912554185e49faff43c685b77764d)