Provide mechanism to pick monetary format from locale info
The display of monetary values varies in different locales but only
parts of the parameters were interpreted by the KMyMoney formatting
function.

This change extracts all necessary information from the current locale
and formats the values accordingly. It is only available on GNU Linux
systems. Windows and MacOS have still to be modified to provide the
necessary information.

Additional unit tests are provided to make sure the formatting works as
expected.

CCBUG: 421915