Initialize width for QML view to the right size
When a user adjusts the size of the Chip-TAN widget, the size is stored
but the view is not adjusted to that size when the dialog is opened the
next time.

This change fixes the problem by setting the view's width accordingly.

BUG: 426243
FIXED-IN: 5.1.1
(cherry picked from commit 1613a08309c9dd29aacdc081496e0420224d7a37)