Provide detailed creation of currencies
The current implementation does not allow to specify all attributes for
currencies but rather uses default values which cannot be changed
afterwards.

This change adds functionality to set all relevant attributes during
creation and also provides a feature to change those which can be
changed after creation of the currency.

BUG: 425935