Do not expose KCMs in application launchers
And they feel really out of place in the application launcher,
it is not clear what is meant with "CSV Importer configuration".
It makes only sense in the context of the settings dialog, where it will
work as before.

BUG: 428776
FIXED-IN: 5.1.1

(cherry picked from commit cdc19a59fb21546938042afd0ef46df63951c058)