Include all transactions in reassignment when merging/deleting payees
The transaction filter skips a few tests when dealing with splits
referencing categories. This causes transactions that only reference
categories but no accounts not to be reported when searching e.g. for
all transactions referencing a specific payee.

Setting a special flag forces to include these tests even on splits
referencing categories. In the given case, this will cause all
transactions to be reported that reference a given set of payees.

BUG: 427519
FIXED-IN: 5.1.1