Fix re-import of online statement file
Due to a duplicate usage of a variable, only a single transaction was
loaded even if the statement file contains more than one statement. This
is a sole debugging feature for developers and has no impact on regular
users.

This change fixes the duplicate usage and importing multiple
transactions now works as expected.