Remove text from account combo box
When selecting a payee that has a default category assigned and then
switching to a payee that does not have a default category assigned
leaves the text of the category as is for the first payee. When the
default category is now enabled, the user has the impression that
the setting is made, but it is not stored.

This change clears the text so that the user has to select the
category in any case which resolves the problem.

BUG: 431248
FIXED-IN: 5.1.2