Fix generation of price entries for securities
In certain circumstances, price entries were generated with the
conversion from currency to security. This caused the automatic price
update to fail.

This change prevents those entries in that it corrects the order and
uses the reciprocal price.

Already existing entries causing the problem are not corrected by this
fix.

See also https://forum.kde.org/viewtopic.php?f=69&t=169077

BUG: 431482
FIXED-IN: 5.1.2