Sort favorite reports on home screen by name
The favorite reports on the home view were shown in the order
the reports were created which is not ideal and confusing.

This change sorts the list of favorite reports on the home
view in alphabetical order.

BUG: 397918
FIXED-IN: 5.1.2

(cherry picked from commit d8c65d062d602e3512d845eaad76a440078cff28)