QIF importer: don't guess account, ask user for it
The account name provided in the QIF file can be ambiguous, so guessing
it by name can return false positives. Moreover, we should not be
creating a new account on the spot if a relevant account cannot be found
. Instead, prepare the statement object so that
MyMoneyStatementReader::selectOrCreateAccount() will show a dialog and
let a user decide.